Criacao de mais alguns endpoints
